requirements.txt
aiocron==1.3
reader==1.6
telethon==1.16.4
PyYAML==5.3.1
